pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "imgcom"
version = "1.5.0"
authors = [{name = "Haris Muntazir", email = "haris.0813.m@gmail.com"}]
description = "A command-line tool to compress images to your desired file size without losing much of the image quality."
readme = "README.md"
license = {file = "LICENSE"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.7"
dependencies = [
    "Pillow",
]

[project.scripts]
imgcom = "imgcom_harismuntazir.compress:main"  

[tool.setuptools]
packages = ["imgcom_harismuntazir"]  
[tool.setuptools.package-data]
"*" = ["LICENSE", "README.md"]
